backend/requirements.txt
Django==2.1.5
djangorestframework==3.9.1
flake8==3.6.0
Pillow==5.3.0
psycopg2==2.7.6.1
factory_boy==2.11.1
